move moving avg module
