+++ internal/movavg/movavg_test.go
@@ -1,4 +1,4 @@
-package blocksync
+package movavg
 
 import (
 	"testing"
